requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
httpx==0.27.2
python-dotenv==1.0.1
openai==1.54.0
pydantic==2.10.0
pydantic-settings==2.6.0
tenacity==8.5.0
structlog==24.4.0

# Database - обновленные версии для Python 3.13
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.11

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0